'burdPuzzleService' gone and renamed
